+++ pkg/ccl/ccl_init.go
@@ -29,7 +29,6 @@ import (
 	_ "github.com/cockroachdb/cockroach/pkg/crosscluster/logical"
 	_ "github.com/cockroachdb/cockroach/pkg/crosscluster/physical"
 	_ "github.com/cockroachdb/cockroach/pkg/crosscluster/producer"
-	_ "github.com/cockroachdb/cockroach/pkg/security/fips"
 	"github.com/cockroachdb/cockroach/pkg/server/license"
 )
